+++ content_service/src/main/java/urdego/api/user/service/UserContentService.java
@@ -1,3 +1,3 @@
-package api.user.service;
+package urdego.api.user.service;
 
 public class UserContentService {}
