+++ content_service/src/main/java/urdego/common/exception/ContentException.java
@@ -1,4 +1,4 @@
-package common.exception;
+package urdego.common.exception;
 
 public abstract class ContentException extends RuntimeException {
     public ContentException(String message) {
