+++ content_service/src/main/java/urdego/common/exception/user/UserContentException.java
@@ -1,6 +1,6 @@
-package common.exception.user;
+package urdego.common.exception.user;
 
-import common.exception.ContentException;
+import urdego.common.exception.ContentException;
 
 public class UserContentException extends ContentException {
     public UserContentException(String message) {
